refactor: move search algorithms to methods directory and update README
